[PORT FROM MAIN] Subject: [PATCH] Do not allow updates to the _data column.
BZ: 137254

Fixes a security hole where applications can update the data location of
voicemail files to point to arbitrary file paths.

Voicemail provider stores the location of the data file in the _data column.
Applications can update this with an arbitrary file path as long as they
have the ADD_VOICEMAIL permission.  Then they can subsequently read that
voicemail and obtain access to the file. This location is generated by the
provider and does not need to be updated by the applications.

Category: Partner Security
Domain: ContactsProvider
Severity: Medium
Origin: internal
Upstream-Candidate: no

Change-Id: I69a4826ccfc12ae6dc147a6d4c35d78e072d1abf
Orig-Change-Id: I1d08a56e7149dc886f96da56d1606d172cce0f55
Signed-off-by: Wu, JunX <junx.wu@intel.com>
Reviewed-on: http://android.intel.com:8080/131052
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>