Integrate Arkham to mainline
BZ: 156819

Arkham is a deeply integrated security container capability for
Android for the purpose of isolation and management in mixed used
cases. Primarily Arkham is intended to enable separation of work and
personal content and capability.

Detailed information about the Arkham requirements and architecture can
be found here: https://sharepoint.amr.ith.intel.com/sites/SDI/Arkham/

This patch modifies the Contacts providers to implement the following
Arkham features:

* Contacts merging between the container contacts and the Android
user who created the container

* Three policies are supported: "no merge", "normal merge" and
"encrypted merge". The policy is set by the container administrator
when the container is created.

* The "no merge" policy does nothing. Thus contacts are isolated
between the primary and container user as per the default Android
policy.

* The "normal merge" policy allows the container contacts to be
merged with the Android user who create the container. When this
policy is active the contacts are stored outside of the container,
thus they are not encrypted. To minize the data leak only part of the
contact information is provided: name, email and phone number. The
actual contacts fieds allowed are contained in a policy set by the
container administrator.This policy has the advantage that contacts
can accessed without having the container opened after rebooting the
phone to decrypt the container data.

* The "ecnrypted merged" policy stores the contact information inside
the container and it is thus encrypted. Because of the the container
contacts are only available if the container is unlock after rebooting
the phone.

In order to minimize the rebase effort, the bulk of the implementation
is done in Extend classes and the original code is only minimally
modified to support hooks for the Extended classes.

In the case of "normal merge" we resolve the container contacts
provider to the primary user contacs provider.

In the case of "encrypted merge" we intercept specific queries and if
we detect they are for container Contacts we communicate with a proxy
inside the container that resolves the request. In order to easily
identify the container contacts we offset container contacts to
"container_user_id * 1000000".

* Container contacts not listed in Primary user when contacts merging is set to NORMAL

* Allow container contact deletion from primary when merge is ENCRYPTED

* Add support for delete container contact.

* Contacts not getting deleted after container is wiped or disabled

Change-Id: Ie58af14169609fb891e769372d8b6f57916d2dd8
Category: feature differentiation
Domain: AOSP.Security-Arkham
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: cactus <cactus@intel.com>